# colmem: improve memory-limiting behavior of the accounting helpers