… pkg/sql/colexec/sort.go
… pkg/sql/colexec/sorttopk.go
… pkg/sql/colfetcher/cfetcher.go
… pkg/sql/colfetcher/vectorized_batch_size_test.go
… pkg/sql/colflow/colrpc/inbox.go
… pkg/sql/colmem/BUILD.bazel
… pkg/sql/colmem/allocator.go
… pkg/sql/colmem/allocator_test.go
… pkg/sql/colmem/reset_maybe_reallocate_test.go
… pkg/sql/opt/exec/execbuilder/testdata/lookup_join_limit