@yuzefovich 2022-08-05
I ran [microbenchmarks](https://gist.github.com/yuzefovich/b5a6e68c45e42a79345ed5962830676b) of `colexec` package, and they mostly show minor speedup. The only concerningly looking result is for `OrderedSynchronizer`, but when running it in isolation, I see no difference, so I think that outlier can be ignored:
```
Switching to HEAD^^^^
+ make bench PKG=./pkg/sql/colexec BENCHTIMEOUT=24h BENCHES=BenchmarkOrderedSynchronizer 'TESTFLAGS=-count 10 -benchmem'
Switching to HEAD
+ make bench PKG=./pkg/sql/colexec BENCHTIMEOUT=24h BENCHES=BenchmarkOrderedSynchronizer 'TESTFLAGS=-count 10 -benchmem'
name                    old time/op    new time/op    delta
OrderedSynchronizer-24    64.6µs ± 8%    61.9µs ± 6%   ~     (p=0.089 n=10+10)

name                    old speed      new speed      delta
OrderedSynchronizer-24   381MB/s ± 8%   398MB/s ± 6%   ~     (p=0.089 n=10+10)

name                    old alloc/op   new alloc/op   delta
OrderedSynchronizer-24     1.00B ± 0%     1.00B ± 0%   ~     (all equal)

name                    old allocs/op  new allocs/op  delta
OrderedSynchronizer-24      0.00           0.00        ~     (all equal)
```

This PR will benefit from an extra set of eyes since I want to backport this to 22.1, so I'll wait for @michae2 to take a look too.