@DrewKimball 2022-08-10
I feel like the 'right' way to solve this in the long term might be to make the various `Set`, `Copy` and `Append` batch and `Vec` methods deterministic in terms of their affect on the memory footprint of the batch. That way, we could check before performing the operation whether it would cause us to exceed the limit and emit immediately if necessary. I guess for `Bytes` columns that would come down to using `make` and `copy` to grow the slice instead of `append`, and other types should be easier.