@yuzefovich 2022-08-10
> I feel like the 'right' way to solve this in the long term might be to make the various Set, Copy and Append methods deterministic in terms of their affect on the memory footprint of the batch. That way, we could check before performing the operation whether it would cause us to exceed the limit and emit immediately if necessary. I guess for Bytes columns that would come down to using make and copy to grow the slice instead of append, and other types should be easier.

It's an interesting idea that'd be worth exploring. We'd need to be conscious about the performance implications though - I'm afraid that introducing those checks would have non-negligible overhead. However, at the same time those conditionals are likely to get very good correct branch prediction most of the time (once the batch has grown to maximum capacity), so maybe the impact would be relatively small. Still we would probably lose the ability to inline most of the `Set` calls.