@yuzefovich 2022-08-10
> I think only bytes columns would require any changes in Set, right? Value types like int would have no change, while pointer types like decimal or datum would just add the size of the object, so should also be deterministic.

Yeah, fair point.